split up TestHelper
