+++ e2e-tests/IIoTPlatform-E2E-Tests/TestHelper.Twin.cs
@@ -1,0 +1,270 @@
+﻿// ------------------------------------------------------------
+//  Copyright (c) Microsoft Corporation.  All rights reserved.
+//  Licensed under the MIT License (MIT). See License.txt in the repo root for license information.
+// ------------------------------------------------------------
+
+namespace IIoTPlatform_E2E_Tests {
+    using IIoTPlatform_E2E_Tests.TestExtensions;
+    using Newtonsoft.Json;
+    using Newtonsoft.Json.Converters;
+    using RestSharp;
+    using System;
+    using System.Collections.Concurrent;
+    using System.Collections.Generic;
+    using System.Dynamic;
+    using System.Linq;
+    using System.Threading;
+    using System.Threading.Tasks;
+    using Xunit;
+
+    internal static partial class TestHelper {
+
+        /// <summary>
+        /// Twin related helper methods
+        /// </summary>
+        public static class Twin {
+
+            /// <summary>
+            /// Equivalent to GetSetOfUniqueNodesAsync
+            /// </summary>
+            /// <param name="context">Shared Context for E2E testing Industrial IoT Platform</param>
+            /// <param name="endpointId">Id of the endpoint as returned by <see cref="Registry_GetEndpoints(IIoTPlatformTestContext)"/></param>
+            /// <param name="nodeId">Id of the parent node or null to browse the root node</param>
+            /// <param name="ct">Cancellation token</param>
+            public static async Task<List<(string NodeId, string NodeClass, bool Children)>> GetBrowseEndpointAsync(
+                    IIoTPlatformTestContext context,
+                    string endpointId,
+                    string nodeId = null,
+                    CancellationToken ct = default) {
+
+                if (string.IsNullOrEmpty(endpointId)) {
+                    context.OutputHelper.WriteLine($"{nameof(endpointId)} is null or empty");
+                    throw new ArgumentNullException(nameof(endpointId));
+                }
+
+                var result = new List<(string NodeId, string NodeClass, bool Children)>();
+                string continuationToken = null;
+
+                do {
+                    var browseResult = await GetBrowseEndpoint_InternalAsync(context, endpointId, nodeId, continuationToken, ct);
+
+                    if (browseResult.results.Count > 0) {
+                        result.AddRange(browseResult.results);
+                    }
+
+                    continuationToken = browseResult.continuationToken;
+                } while (continuationToken != null);
+
+                return result;
+            }
+
+            /// <summary>
+            /// Calls a GET twin browse with the given <paramref name="endpointId"/>
+            /// </summary>
+            /// <param name="context">Shared Context for E2E testing Industrial IoT Platform</param>
+            /// <param name="endpointId">Id of the endpoint as returned by <see cref="Registry_GetEndpoints(IIoTPlatformTestContext)"/></param>
+            /// <param name="nodeId">Id of the parent node or null to browse the root node</param>
+            /// <param name="continuationToken">Continuation token from the previous call, or null</param>
+            /// <param name="ct">Cancellation token</param>
+            private static async Task<(List<(string NodeId, string NodeClass, bool Children)> results, string continuationToken)> GetBrowseEndpoint_InternalAsync(
+                    IIoTPlatformTestContext context,
+                    string endpointId,
+                    string nodeId = null,
+                    string continuationToken = null,
+                    CancellationToken ct = default) {
+
+                var accessToken = await GetTokenAsync(context, ct).ConfigureAwait(false);
+                var client = new RestClient(context.IIoTPlatformConfigHubConfig.BaseUrl) { Timeout = TestConstants.DefaultTimeoutInMilliseconds };
+
+                var request = new RestRequest(Method.GET);
+                request.AddHeader(TestConstants.HttpHeaderNames.Authorization, accessToken);
+
+                if (continuationToken == null) {
+                    request.Resource = $"twin/v2/browse/{endpointId}";
+
+                    if (!string.IsNullOrEmpty(nodeId)) {
+                        request.AddQueryParameter("nodeId", nodeId);
+                    }
+                }
+                else {
+                    request.Resource = $"twin/v2/browse/{endpointId}/next";
+                    request.AddQueryParameter("continuationToken", continuationToken);
+                }
+
+                var response = await client.ExecuteAsync(request, ct).ConfigureAwait(false);
+
+                Assert.NotNull(response);
+                if (!response.IsSuccessful) {
+                    context.OutputHelper.WriteLine($"StatusCode: {response.StatusCode}");
+                    context.OutputHelper.WriteLine($"ErrorMessage: {response.ErrorMessage}");
+                    Assert.True(response.IsSuccessful, "GET twin/v2/browse/{endpointId} failed!");
+                }
+
+                dynamic json = JsonConvert.DeserializeObject<ExpandoObject>(response.Content, new ExpandoObjectConverter());
+
+                Assert.True(HasProperty(json, "references"), "GET twin/v2/browse/{endpointId} response has no items");
+                Assert.False(json.references == null, "GET twin/v2/browse/{endpointId} response references property is null");
+
+                var result = new List<(string NodeId, string NodeClass, bool Children)>();
+
+                foreach (var node in json.references) {
+                    result.Add(
+                        (
+                            node.target?.nodeId?.ToString(),
+                            node.target?.nodeClass?.ToString(),
+                            string.Equals(node.target?.children?.ToString(), "true", StringComparison.OrdinalIgnoreCase)));
+                }
+
+                var responseContinuationToken = HasProperty(json, "continuationToken") ? json.continuationToken : null;
+
+                return (results: result, continuationToken: responseContinuationToken);
+            }
+
+            /// <summary>
+            /// Equivalent to recursive calling GetSetOfUniqueNodesAsync to get the whole hierarchy of nodes
+            /// </summary>
+            /// <param name="context">Shared Context for E2E testing Industrial IoT Platform</param>
+            /// <param name="endpointId">Id of the endpoint as returned by <see cref="Registry_GetEndpoints(IIoTPlatformTestContext)"/></param>
+            /// <param name="nodeClass">Class of the node to filter to or null for no filtering</param>
+            /// <param name="nodeId">Id of the parent node or null to browse the root node</param>
+            /// <param name="ct">Cancellation token</param>
+            public static async Task<List<(string NodeId, string NodeClass, bool Children)>> GetBrowseEndpointRecursiveAsync(
+                    IIoTPlatformTestContext context,
+                    string endpointId,
+                    string nodeClass = null,
+                    string nodeId = null,
+                    CancellationToken ct = default) {
+
+                if (string.IsNullOrEmpty(endpointId)) {
+                    context.OutputHelper.WriteLine($"{nameof(endpointId)} is null or empty");
+                    throw new ArgumentNullException(nameof(endpointId));
+                }
+
+                var nodes = new ConcurrentBag<(string NodeId, string NodeClass, bool Children)>();
+
+                await GetBrowseEndpointRecursiveCollectResultsAsync(context, endpointId, nodes, nodeId, ct);
+
+                return nodes.Where(n => string.Equals(nodeClass, n.NodeClass, StringComparison.OrdinalIgnoreCase)).ToList();
+            }
+
+            /// <summary>
+            /// Collects all nodes recursively avoiding circular references between nodes
+            /// </summary>
+            /// <param name="context">Shared Context for E2E testing Industrial IoT Platform</param>
+            /// <param name="endpointId">Id of the endpoint as returned by <see cref="Registry_GetEndpoints(IIoTPlatformTestContext)"/></param>
+            /// <param name="nodes">Collection of nodes found</param>
+            /// <param name="nodeId">Id of the parent node or null to browse the root node</param>
+            /// <param name="ct">Cancellation token</param>
+            private static async Task GetBrowseEndpointRecursiveCollectResultsAsync(
+                    IIoTPlatformTestContext context,
+                    string endpointId,
+                    ConcurrentBag<(string NodeId, string NodeClass, bool Children)> nodes,
+                    string nodeId = null,
+                    CancellationToken ct = default) {
+
+                var currentNodes = await Twin.GetBrowseEndpointAsync(context, endpointId, nodeId).ConfigureAwait(false);
+
+                foreach (var node in currentNodes) {
+                    ct.ThrowIfCancellationRequested();
+
+                    if (nodes.Any(n => string.Equals(n.NodeId, node.NodeId))) {
+                        continue;
+                    }
+
+                    nodes.Add(node);
+
+                    if (node.Children) {
+                        await GetBrowseEndpointRecursiveCollectResultsAsync(
+                            context,
+                            endpointId,
+                            nodes,
+                            node.NodeId,
+                            ct).ConfigureAwait(false);
+                    }
+                }
+            }
+
+            /// <summary>
+            /// Gets method metadata
+            /// </summary>
+            /// <param name="context">Shared Context for E2E testing Industrial IoT Platform</param>
+            /// <param name="endpointId">Id of the endpoint as returned by <see cref="Registry_GetEndpoints(IIoTPlatformTestContext)"/></param>
+            /// <param name="methodId">Id of the method to get the metadata of</param>
+            /// <param name="ct">Cancellation token</param>
+            public static async Task<dynamic> GetMethodMetadataAsync(
+            IIoTPlatformTestContext context,
+                string endpointId,
+                string methodId = null,
+                CancellationToken ct = default) {
+
+                var accessToken = await GetTokenAsync(context, ct).ConfigureAwait(false);
+                var client = new RestClient(context.IIoTPlatformConfigHubConfig.BaseUrl) { Timeout = TestConstants.DefaultTimeoutInMilliseconds };
+
+                var request = new RestRequest(Method.POST);
+                request.AddHeader(TestConstants.HttpHeaderNames.Authorization, accessToken);
+
+                request.Resource = $"twin/v2/call/{endpointId}/metadata";
+
+                var body = new {
+                    methodId,
+                    header = new {
+                        diagnostics = new {
+                            level = "Verbose"
+                        }
+                    }
+                };
+
+                request.AddJsonBody(JsonConvert.SerializeObject(body));
+
+                var response = await client.ExecuteAsync(request, ct).ConfigureAwait(false);
+                dynamic json = JsonConvert.DeserializeObject<ExpandoObject>(response.Content, new ExpandoObjectConverter());
+
+                return json;
+            }
+
+            /// <summary>
+            /// Calls a method
+            /// </summary>
+            /// <param name="context">Shared Context for E2E testing Industrial IoT Platform</param>
+            /// <param name="endpointId">Id of the endpoint as returned by <see cref="Registry_GetEndpoints(IIoTPlatformTestContext)"/></param>
+            /// <param name="methodId">Id of the method to call</param>
+            /// <param name="objectId">Object ID</param>
+            /// <param name="arguments">Method arguments</param>
+            /// <param name="ct">Cancellation token</param>
+            public static async Task<dynamic> CallMethodAsync(
+                IIoTPlatformTestContext context,
+                    string endpointId,
+                    string methodId,
+                    string objectId,
+                    List<object> arguments,
+                    CancellationToken ct = default) {
+
+                var accessToken = await GetTokenAsync(context, ct).ConfigureAwait(false);
+                var client = new RestClient(context.IIoTPlatformConfigHubConfig.BaseUrl) { Timeout = TestConstants.DefaultTimeoutInMilliseconds };
+
+                var request = new RestRequest(Method.POST);
+                request.AddHeader(TestConstants.HttpHeaderNames.Authorization, accessToken);
+
+                request.Resource = $"twin/v2/call/{endpointId}";
+
+                var body = new {
+                    methodId,
+                    objectId,
+                    arguments,
+                    header = new {
+                        diagnostics = new {
+                            level = "Verbose"
+                        }
+                    }
+                };
+
+                request.AddJsonBody(JsonConvert.SerializeObject(body));
+
+                var response = await client.ExecuteAsync(request, ct).ConfigureAwait(false);
+                dynamic json = JsonConvert.DeserializeObject<ExpandoObject>(response.Content, new ExpandoObjectConverter());
+
+                return json;
+            }
+        }
+    }
+}
